Replace h1 small .show-for-large-up with a media query trick
.show-for-large-up inherits the display of the parent which is h1
so this causes it to be displayed as a block which does not look
good on large screens. Instead we use a media query and add the
h1 small text (which was hidden on smaller devices) but we tell
the smaller devices to display it as a block (which looks ok).